Bugfix: Integer Casting entfernt.
Ich hatte vergessen den int-Cast von TweetID und AutorID zu entfernen.
Jetzt sollte alles funktionieren und die IDs stehen richtig in der DB.

+++ TwitterMonitor/src/de/htwsaar/twitter/Tweet.java
@@ -24,8 +24,8 @@ public class Tweet {
 	}
 
 	public Tweet(Status status) {
-		this.tweetId = (int) status.getId();
-		this.authorId = (int) status.getUser().getId();
+		this.tweetId = status.getId();
+		this.authorId = status.getUser().getId();
 		this.text = status.getText();
 		this.createdAt = status.getCreatedAt();
 		this.place = status.getPlace().getCountry();
